# public/api/openapi.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Laravel Chat",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://laravel-chat.test/api",
    "description" : "Laravel chat project"
  } ],
  "paths" : {
...
    "/user/conversation/{conversationId}" : {
      "post" : {
        "tags" : [ "Conversation" ],
        "summary" : "update conversation name and photo",
        "operationId" : "updateConversation",
        "parameters" : [ {
          "name" : "conversationId",
          "in" : "path",
          "description" : "ID of conversation to update",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
          {
            "name": "custom_name",
            "in": "query",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string"
            }
          }],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "custom_photo": {
                    "type":"file",
                    "format": "base64"
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Conversation"
                }
              }
            }
          }
        },
        "security" : [ {
          "Authorisation" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "Conversation" ],
        "summary" : "Deletes a conversation",
        "operationId" : "deleteConversation",
        "parameters" : [ {
          "name" : "conversationId",
          "in" : "path",
          "description" : "Conversation id to delete",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Conversation deleted successfully"
          },
          "404" : {
            "description" : "Conversation not found"
          }
        },
        "security" : [ {
          "Authorisation" : [ ]
        } ]
      }
    },
    "/user/conversation/{conversationId}/members" : {
      "get" : {
        "tags" : [ "Conversation" ],
        "summary" : "Conversation members",
        "operationId" : "conversationMembers",
        "parameters" : [ {
          "name" : "conversationId",
          "in" : "path",
          "description" : "Conversation id which will be used to get members ",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "integer",
                    "format" : "int32"
                  }
                }
              }
            }
          }
        },
        "security" : [ {
          "Authorisation" : [ ]
        } ]
      }
    },
    "/user/conversation/{conversationId}/leave" : {
      "get" : {
        "tags" : [ "Conversation" ],
        "summary" : "Leave conversation",
        "operationId" : "leaveConversation",
        "parameters" : [ {
          "name" : "conversationId",
          "in" : "path",
          "description" : "Conversation Id ",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "integer",
                    "format" : "int32"
                  }
                }
              }
            }
          }
        },
        "security" : [ {
          "Authorisation" : [ ]
        } ]
      }
    },
...
    "/user/photos/{photoId}" : {
      "delete" : {
        "tags" : [ "Photo" ],
        "summary" : "Delete photo by ID",
        "description" : "For valid response try integer IDs with positive integer value.\\ \\ Negative or non-integer values will generate API errors",
        "operationId" : "deletePhoto",
        "parameters" : [ {
          "name" : "photoId",
          "in" : "path",
          "description" : "ID of the photo that needs to be deleted",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "minimum" : 1,
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "description" : "Photo not found"
          }
        },
        "security" : [ {
          "Authorisation" : [ ]
        } ]
      },
      "get" : {
        "tags" : [ "Photo" ],
        "summary" : "Set profile photo",
        "operationId" : "profilePhoto",
        "parameters" : [ {
          "name" : "photoId",
          "in" : "path",
          "description" : "ID of the photo that will be setted as profile photo",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "minimum" : 1,
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Profile photo updated successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Photo"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Photo"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid Photo"
          }
        },
        "security" : [ {
          "Authorisation" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Photo" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "user_id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "photo" : {
            "type" : "string"
          }
        },
        "xml" : {
          "name" : "Order"
        }
      },
      "Conversation" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "custom_name" : {
            "type" : "string"
          },
          "custom_photo" : {
            "type" : "string"
          }
        },
        "xml" : {
          "name" : "Category"
        }
      },
...
    },
...
    "securitySchemes" : {
      "Authorisation" : {
        "type" : "http",
        "scheme": "bearer"
      }
    }
  }
}